Beta 1.0
World 3 is now complete.
- Overhauled Tank movement.
- Overhauled World Overview.
- Overhauled UI.
- Improved BossBar.
- Various Bugfixes.

+++ Assets/Scripts/Classes/SaveV1.cs
@@ -48,7 +48,7 @@ public class SaveV1 : ISaveGame {
 		SaveGame.Save();
 	}
 
-	public void UpdateSaveSlot(byte slot, ulong levelId, byte lives, short score, float time) {
+	public void UpdateSaveSlot(byte slot, ulong levelId, byte lives, int score, float time) {
 		switch(slot) {
 			case 0:
 				saveSlot1.levelId = levelId;
